+++ src/agent/src/Approval/Exception/ToolApprovalException.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\AI\Agent\Approval\Exception;
+
+use Symfony\AI\Agent\Exception\RuntimeException;
+
+/**
+ * @author Saiful Islam <saif012@gmail.com>
+ */
+class ToolApprovalException extends RuntimeException implements ExceptionInterface
+{
+}